Update faq.md to have updated info
theres 2 ways to get banned. cheat online, or eShop fraud. so i updated it for eshop fraud
for the haxchi offline, indexiine is a thing that lets you use the browser exploit offline. so you dont really need haxchi
and also, makes it read that it supports the latest firmware (5.5.5)