Tiramisu SD Preparation: Cover versions.json
Step 2 of the Instructions section asks to copy the Tiramisu .zip file's contents to the root of an SD card. But the SD Card Layout section lacked any mention the zip file's top-level versions.json file.

This adds versions.json to the end of the SD Card Layout's file tree.