index: add hack to restore docsify behaviour
This attempts to restore the old behaviour of docsify.js, where it uses
hash parameters to navigate the site. Since this isn't the case in
VitePress, all the old links to the guide broke, so check it on load.

This works on a best-effort basis; if the corresponding page exists
(such as `#/aroma/getting-started` -> `/aroma/getting-started`), it
should navigate properly, but if the corresponding page doesn't exist,
it will 404.